[lesson-1] create employee use case, refactor code

+++ internal/adapter/in_memory_db/employee_db/find.go
@@ -5,7 +5,7 @@ import (
 	"warehouse_project/internal/domain/model"
 )
 
-func (er *EmployeeRepo) FindEmployee(id int) (*model.Employee, error) {
+func (er *EmployeeRepo) FindEmployee(id int64) (*model.Employee, error) {
 	employee, exists := er.employees[id]
 	if !exists {
 		return nil, fmt.Errorf("Employee not found")
