+++ internal/adapter/in_memory_db/order_db/find.go
@@ -2,7 +2,7 @@ package order_db
 
 import "warehouse_project/internal/domain/model"
 
-func (or *OrderRepo) FindOrder(id int) (*model.Order, bool) {
+func (or *OrderRepo) FindOrder(id int64) (*model.Order, bool) {
 	order, exists := or.orders[id]
 	return order, exists
 }
